Fix errors when scanning files > 4G
This commit resolves https://bugzilla.clamav.net/show_bug.cgi?id=12673

Changes in 0.103 to order of operations for creating fmaps and
performaing hashes of fmaps resulted errors when scanning files that are
4096M and a different (but related) error when scanning files > 4096M.
This is despite the fact that scanning is supposed to be limited to
--max-scansize (MaxScanSize) and was also apparently limited to
INT_MAX - 2 (aka ~1.999999G) back in 2014 to alleviate reported crashes
for a few large file formats.
(see https://bugzilla.clamav.net/show_bug.cgi?id=10960)
This last limitation was not documented, so I added it to the sample
clamd.conf.

Anyways, the main issue is that the fmap module was using "unsigned int"
and was then enforcing a limitation (verbose error messages) when that
a map length exceeded the capapacity of an unsigned int. This commit
switches the associated variables over to uint64_t, and while fmaps are
still limited to size_t in other places, the fmap module will at least
work with files > 4G on 64bit systems.

In testing this, I found that the time to hash a file, particularly when
hashing a file on an NTFS partition from Linux was really slow because
we were hashing in FILEBUFF chunks (about 8K) at a time.  Increasing
this to 10MB chunks speeds up scanning of large files.

Finally, now that hashing is performed immediately when an fmap is
created for a file, hashing of files larger than max-scansize was
occuring. This commit adds checks to bail out early if the file size
exceeds the maximum before creating an fmap. It will alert with the
Heuristics.Limits.Exceeded name if the heuristic is enabled.

Also fixed CheckFmapFeatures.cmake module that detects if
sysconf(_SC_PAGESIZE) is available.

+++ libclamav/fmap.h
@@ -46,12 +46,12 @@ struct cl_fmap {
 
     /* internal */
     time_t mtime;
-    unsigned int pages;
+    uint64_t pages;
     uint64_t pgsz;
-    unsigned int paged;
-    unsigned short aging;
-    unsigned short dont_cache_flag;
-    unsigned short handle_is_fd;
+    uint64_t paged;
+    uint16_t aging;
+    uint16_t dont_cache_flag;
+    uint16_t handle_is_fd;
 
     /* memory interface */
     const void *data;
@@ -81,7 +81,7 @@ struct cl_fmap {
     HANDLE mh;
 #endif
     unsigned char maphash[16];
-    uint32_t *bitmap;
+    uint64_t *bitmap;
     char *name;
 };
